# fix: optimise Is* and As* to reduce allocations