This rework the `Is` and `As` helper functions to reduce or remove memory allocation.

## Comparison

### Is*

| Function | Before | After | Speed Improvement | Memory Reduction | Allocation Reduction |
|----------|--------|-------|-------------------|------------------|---------------------|
| `IsPermissionDenied` | 25.69 ns/op, 16 B/op, 1 allocs/op | 6.67 ns/op, 0 B/op, 0 allocs/op | **3.8x faster** | **100%** | **100%** |
| `IsUnauthenticated` | 25.11 ns/op, 16 B/op, 1 allocs/op | 6.46 ns/op, 0 B/op, 0 allocs/op | **3.9x faster** | **100%** | **100%** |
| `IsNotFound` | 24.97 ns/op, 16 B/op, 1 allocs/op | 6.44 ns/op, 0 B/op, 0 allocs/op | **3.9x faster** | **100%** | **100%** |
| `IsBad` | 29.68 ns/op, 48 B/op, 1 allocs/op | 7.52 ns/op, 0 B/op, 0 allocs/op | **3.9x faster** | **100%** | **100%** |
| `IsFailedPrecondition` | 29.66 ns/op, 48 B/op, 1 allocs/op | 6.64 ns/op, 0 B/op, 0 allocs/op | **4.5x faster** | **100%** | **100%** |
| `IsAborted` | 30.16 ns/op, 48 B/op, 1 allocs/op | 6.45 ns/op, 0 B/op, 0 allocs/op | **4.7x faster** | **100%** | **100%** |
| `IsUnavailable` | 27.14 ns/op, 24 B/op, 1 allocs/op | 6.56 ns/op, 0 B/op, 0 allocs/op | **4.1x faster** | **100%** | **100%** |
| `IsResourceExhausted` | 31.34 ns/op, 48 B/op, 1 allocs/op | 6.46 ns/op, 0 B/op, 0 allocs/op | **4.9x faster** | **100%** | **100%** |
| `IsUnimplemented` | 25.20 ns/op, 16 B/op, 1 allocs/op | 6.51 ns/op, 0 B/op, 0 allocs/op | **3.9x faster** | **100%** | **100%** |

### As*

| Function | Before | After | Speed Improvement | Memory Reduction | Allocation Reduction |
|----------|--------|-------|-------------------|------------------|---------------------|
| `AsPermissionDenied` | 75.67 ns/op, 24 B/op, 2 allocs/op | 55.02 ns/op, 8 B/op, 1 allocs/op | **1.4x faster** | **67%** | **50%** |
| `AsUnauthenticated` | 77.50 ns/op, 24 B/op, 2 allocs/op | 55.34 ns/op, 8 B/op, 1 allocs/op | **1.4x faster** | **67%** | **50%** |
| `AsNotFound` | 75.80 ns/op, 24 B/op, 2 allocs/op | 55.36 ns/op, 8 B/op, 1 allocs/op | **1.4x faster** | **67%** | **50%** |
| `AsBad` | 83.31 ns/op, 56 B/op, 2 allocs/op | 55.30 ns/op, 8 B/op, 1 allocs/op | **1.5x faster** | **86%** | **50%** |
| `AsFailedPrecondition` | 82.36 ns/op, 56 B/op, 2 allocs/op | 55.64 ns/op, 8 B/op, 1 allocs/op | **1.5x faster** | **86%** | **50%** |
| `AsAborted` | 83.82 ns/op, 56 B/op, 2 allocs/op | 55.21 ns/op, 8 B/op, 1 allocs/op | **1.5x faster** | **86%** | **50%** |
| `AsUnavailable` | 85.54 ns/op, 32 B/op, 2 allocs/op | 56.48 ns/op, 8 B/op, 1 allocs/op | **1.5x faster** | **75%** | **50%** |
| `AsResourceExhausted` | 82.19 ns/op, 56 B/op, 2 allocs/op | 55.30 ns/op, 8 B/op, 1 allocs/op | **1.5x faster** | **86%** | **50%** |
| `AsUnimplemented` | 75.35 ns/op, 24 B/op, 2 allocs/op | 55.30 ns/op, 8 B/op, 1 allocs/op | **1.4x faster** | **67%** | **50%** |

## Technical Details

### Is* Functions Optimization
```go
// Before
func IsPermissionDenied(err error) bool {
    return errors.Is(err, &PermissionFailure{})  // Allocates every call
}

// After  
func IsPermissionDenied(err error) bool {
    return errors.Is(err, permissionFailureInstance)  // Reuses singleton
}
```

### As* Functions Optimization
```go
// Before
func AsPermissionDenied(err error) (*PermissionFailure, bool) {
    e := &PermissionFailure{}  // Unnecessary allocation
    if errors.As(err, &e) {
        return e, true
    }
    return nil, false
}

// After
func AsPermissionDenied(err error) (*PermissionFailure, bool) {
    var e *PermissionFailure  // No allocation, starts as nil
    if errors.As(err, &e) {
        return e, true
    }
    return nil, false
}
```